ueh-web-dev/backend/requirements.txt
Package                       Version
----------------------------- -----------
aiohttp                       3.9.5
aiosignal                     1.3.1
annotated-types               0.7.0
anyio                         4.4.0
argon2-cffi                   23.1.0
argon2-cffi-bindings          21.2.0
asgiref                       3.7.2
async-timeout                 4.0.3
attrs                         23.2.0
backports.zoneinfo            0.2.1
bcrypt                        4.1.2
certifi                       2024.6.2
cffi                          1.16.0
charset-normalizer            3.3.2
click                         8.1.7
cmake                         3.29.0.1
colorama                      0.4.6
contourpy                     1.1.1
cycler                        0.12.1
datasets                      2.20.0
diff-match-patch              20230430
dill                          0.3.8
distro                        1.9.0
dj-database-url               2.2.0
Django                        4.2.10
django-bootstrap3             23.6
django-braces                 1.15.0
django-ckeditor               6.6.1
django-ckeditor-5             0.2.12
django-cors-headers           4.3.1
django-crispy-forms           2.1
django-debug-toolbar          4.3.0
django-heroku                 0.3.1
django-import-export          4.0.0
django-jazzmin                3.0.0
django-js-asset               2.2.0
django-mathfilters            1.0.0
django-taggit                 5.0.1
djangorestframework           3.14.0
djangorestframework-simplejwt 5.3.1
dlib                          19.24.4
docker-pycreds                0.4.0
et-xmlfile                    1.1.0
exceptiongroup                1.2.1
face-recognition              1.3.0
face-recognition-models       0.3.0
filelock                      3.13.3
fonttools                     4.50.0
frozenlist                    1.4.1
fsspec                        2024.3.1
gitdb                         4.0.11
GitPython                     3.1.43
gunicorn                      22.0.0
h11                           0.14.0
httpcore                      1.0.5
httpx                         0.27.0
huggingface-hub               0.23.4
idna                          3.7
importlib_resources           6.4.0
Jinja2                        3.1.3
kiwisolver                    1.4.5
MarkupSafe                    2.1.5
matplotlib                    3.7.5
misaka                        2.1.1
mpmath                        1.3.0
multidict                     6.0.5
multiprocess                  0.70.16
networkx                      3.1
numpy                         1.24.4
openai                        1.34.0
opencv-python                 4.9.0.80
openpyxl                      3.1.4
packaging                     24.0
pandas                        2.0.3
pillow                        10.2.0
pip                           23.3.1
platformdirs                  4.2.2
protobuf                      5.27.2
psutil                        6.0.0
psycopg2                      2.9.9
psycopg2-binary               2.9.9
pyarrow                       16.1.0
pyarrow-hotfix                0.6
pycparser                     2.21
pydantic                      2.7.4
pydantic_core                 2.18.4
PyJWT                         2.8.0
pyparsing                     3.1.2
python-dateutil               2.9.0.post0
python-dotenv                 1.0.1
pytz                          2024.1
PyYAML                        6.0.1
regex                         2024.5.15
requests                      2.32.3
safetensors                   0.4.3
sentry-sdk                    2.7.0
setproctitle                  1.3.3
setuptools                    68.2.2
six                           1.16.0
smmap                         5.0.1
sniffio                       1.3.1
sqlparse                      0.4.4
sympy                         1.12
tablib                        3.5.0
tiktoken                      0.7.0
tokenizers                    0.19.1
torch                         2.2.2
torchvision                   0.17.2
tqdm                          4.66.4
transformers                  4.41.2
typing_extensions             4.9.0
tzdata                        2024.1
urllib3                       2.2.2
validate-email                1.3
wandb                         0.17.3
wheel                         0.41.2
whitenoise                    6.6.0
xxhash                        3.4.1
yarl                          1.9.4
zipp                          3.18.1
